WIP: flesh out `wradlib.io` submodules, remove deprecated functions, fix tests/notebooks
