+++ wradlib/io/hdf.py
@@ -1,0 +1,441 @@
+#!/usr/bin/env python
+# Copyright (c) 2011-2017, wradlib developers.
+# Distributed under the MIT License. See LICENSE.txt for more info.
+
+"""
+HDF Data I/O
+^^^^^^^^^^^^
+.. autosummary::
+   :nosignatures:
+   :toctree: generated/
+
+   read_generic_hdf5
+   read_OPERA_hdf5
+   read_GAMIC_hdf5
+   to_hdf5
+   from_hdf5
+"""
+
+# standard libraries
+from __future__ import absolute_import
+import sys
+
+# site packages
+import h5py
+import numpy as np
+
+
+def browse_hdf5_group(grp):
+    """Browses one hdf5 file level
+    """
+    pass
+
+
+def read_generic_hdf5(fname):
+    """Reads hdf5 files according to their structure
+
+    In contrast to other file readers under :meth:`wradlib.io`, this function
+    will *not* return a two item tuple with (data, metadata). Instead, this
+    function returns ONE dictionary that contains all the file contents - both
+    data and metadata. The keys of the output dictionary conform to the
+    Group/Subgroup directory branches of the original file.
+
+    Parameters
+    ----------
+    fname : string
+        a hdf5 file path
+
+    Returns
+    -------
+    output : dict
+        a dictionary that contains both data and metadata according to the
+        original hdf5 file structure
+
+    Examples
+    --------
+    See :ref:`notebooks/fileio/wradlib_radar_formats.ipynb#Generic-HDF5`.
+    """
+    f = h5py.File(fname, "r")
+    fcontent = {}
+
+    def filldict(x, y):
+        # create a new container
+        tmp = {}
+        # add attributes if present
+        if len(y.attrs) > 0:
+            tmp['attrs'] = dict(y.attrs)
+        # add data if it is a dataset
+        if isinstance(y, h5py.Dataset):
+            tmp['data'] = np.array(y)
+        # only add to the dictionary, if we have something meaningful to add
+        if tmp != {}:
+            fcontent[x] = tmp
+
+    f.visititems(filldict)
+
+    f.close()
+
+    return fcontent
+
+
+def read_OPERA_hdf5(fname):
+    """Reads hdf5 files according to OPERA conventions
+
+    Please refer to the OPERA data model documentation :cite:`OPERA-data-model`
+    in order to understand how an hdf5 file is organized that conforms to the
+    OPERA ODIM_H5 conventions.
+
+    In contrast to other file readers under :meth:`wradlib.io`, this function
+    will *not* return a two item tuple with (data, metadata). Instead, this
+    function returns ONE dictionary that contains all the file contents - both
+    data and metadata. The keys of the output dictionary conform to the
+    Group/Subgroup directory branches of the original file.
+    If the end member of a branch (or path) is "data", then the corresponding
+    item of output dictionary is a numpy array with actual data.
+
+    Any other end member (either *how*, *where*,
+    and *what*) will contain the meta information applying to the corresponding
+    level of the file hierarchy.
+
+    Parameters
+    ----------
+    fname : string
+        a hdf5 file path
+
+    Returns
+    -------
+    output : dict
+        a dictionary that contains both data and metadata according to the
+        original hdf5 file structure
+    """
+    f = h5py.File(fname, "r")
+
+    # now we browse through all Groups and Datasets and store the info in one
+    # dictionary
+    fcontent = {}
+
+    def filldict(x, y):
+        if isinstance(y, h5py.Group):
+            if len(y.attrs) > 0:
+                fcontent[x] = dict(y.attrs)
+        elif isinstance(y, h5py.Dataset):
+            fcontent[x] = np.array(y)
+
+    f.visititems(filldict)
+
+    f.close()
+
+    return fcontent
+
+
+def read_gamic_scan_attributes(scan, scan_type):
+    """Read attributes from one particular scan from a GAMIC hdf5 file
+
+    Provided by courtesy of Kai Muehlbauer (University of Bonn).
+
+    Parameters
+    ----------
+    scan : object
+        scan object from hdf5 file
+    scan_type : string
+        "PVOL" (plan position indicator) or "RHI" (range height indicator)
+
+    Returns
+    -------
+    sattrs : dict
+        dictionary of scan attributes
+    """
+
+    # global zero_index, el, az
+
+    # placeholder for attributes
+    sattrs = {}
+
+    # link to scans 'how' hdf5 group
+    sg1 = scan['how']
+
+    # get scan attributes
+    for attrname in list(sg1.attrs):
+        sattrs[attrname] = sg1.attrs.get(attrname)
+    sattrs['bin_range'] = sattrs['range_step'] * sattrs['range_samples']
+
+    # get scan header
+    ray_header = scan['ray_header']
+
+    # az, el, zero_index for PPI scans
+    if scan_type == 'PVOL':
+        azi_start = ray_header['azimuth_start']
+        azi_stop = ray_header['azimuth_stop']
+        # Azimuth corresponding to 1st ray
+        zero_index = np.where(azi_stop < azi_start)
+        azi_stop[zero_index[0]] += 360
+        zero_index = zero_index[0] + 1
+        az = (azi_start + azi_stop) / 2
+        az = np.roll(az, -zero_index, axis=0)
+        az = np.round(az, 1)
+        el = sg1.attrs.get('elevation')
+
+    # az, el, zero_index for RHI scans
+    if scan_type == 'RHI':
+        ele_start = np.round(ray_header['elevation_start'], 1)
+        ele_stop = np.round(ray_header['elevation_stop'], 1)
+        angle_step = np.round(sattrs['angle_step'], 1)
+        angle_step = int(np.round(sattrs['ele_stop'], 1) / angle_step)
+        # Elevation corresponding to 1st ray
+        if ele_start[0] < 0:
+            ele_start = ele_start[1:]
+            ele_stop = ele_stop[1:]
+        zero_index = np.where(ele_stop > ele_start)
+        zero_index = zero_index[0]  # - 1
+        el = (ele_start + ele_stop) / 2
+        el = np.round(el, 1)
+        el = el[-angle_step:]
+
+        az = sg1.attrs.get('azimuth')
+
+    # save zero_index (first ray) to scan attributes
+    sattrs['zero_index'] = zero_index[0]
+
+    # create range array
+    r = np.arange(sattrs['bin_range'],
+                  sattrs['bin_range'] * sattrs['bin_count'] +
+                  sattrs['bin_range'], sattrs['bin_range'])
+
+    # save variables to scan attributes
+    sattrs['az'] = az
+    sattrs['el'] = el
+    sattrs['r'] = r
+    sattrs['Time'] = sattrs.pop('timestamp')
+    sattrs['max_range'] = r[-1]
+
+    return sattrs
+
+
+def read_gamic_scan(scan, scan_type, wanted_moments):
+    """Read data from one particular scan from GAMIC hdf5 file
+
+    Provided by courtesy of Kai Muehlbauer (University of Bonn).
+
+    Parameters
+    ----------
+    scan : object
+        scan object from hdf5 file
+    scan_type : string
+        "PVOL" (plan position indicator) or "RHI" (range height indicator)
+    wanted_moments : strings
+        sequence of strings containing upper case names of moment(s) to
+        be returned
+
+    Returns
+    -------
+    data : dict
+        dictionary of moment data (numpy arrays)
+    sattrs : dict
+        dictionary of scan attributes
+    """
+
+    # placeholder for data and attrs
+    data = {}
+    sattrs = {}
+
+    # try to read wanted moments
+    for mom in list(scan):
+        if 'moment' in mom:
+            data1 = {}
+            sg2 = scan[mom]
+            actual_moment = sg2.attrs.get('moment').decode().upper()
+            if (actual_moment in wanted_moments) or (wanted_moments == 'all'):
+                # read attributes only once
+                if not sattrs:
+                    sattrs = read_gamic_scan_attributes(scan, scan_type)
+                mdata = sg2[...]
+                dyn_range_max = sg2.attrs.get('dyn_range_max')
+                dyn_range_min = sg2.attrs.get('dyn_range_min')
+                bin_format = sg2.attrs.get('format').decode()
+                if bin_format == 'UV8':
+                    div = 256.0
+                else:
+                    div = 65536.0
+                mdata = (dyn_range_min + mdata *
+                         (dyn_range_max - dyn_range_min) / div)
+
+                if scan_type == 'PVOL':
+                    # rotate accordingly
+                    mdata = np.roll(mdata, -1 * sattrs['zero_index'], axis=0)
+
+                if scan_type == 'RHI':
+                    # remove first zero angles
+                    sdiff = mdata.shape[0] - sattrs['el'].shape[0]
+                    mdata = mdata[sdiff:, :]
+
+                data1['data'] = mdata
+                data1['dyn_range_max'] = dyn_range_max
+                data1['dyn_range_min'] = dyn_range_min
+                data[actual_moment] = data1
+
+    return data, sattrs
+
+
+def read_GAMIC_hdf5(filename, wanted_elevations=None, wanted_moments=None):
+    """Data reader for hdf5 files produced by the commercial
+    GAMIC Enigma V3 MURAN software
+
+    Provided by courtesy of Kai Muehlbauer (University of Bonn). See GAMIC
+    homepage for further info (http://www.gamic.com).
+
+    Parameters
+    ----------
+    filename : string
+        path of the gamic hdf5 file
+    wanted_elevations : strings
+        sequence of strings of elevation_angle(s) of scan (only needed for PPI)
+    wanted_moments : strings
+        sequence of strings of moment name(s)
+
+    Returns
+    -------
+    data : dict
+        dictionary of scan and moment data (numpy arrays)
+    attrs : dict
+        dictionary of attributes
+
+    Examples
+    --------
+    See :ref:`notebooks/fileio/wradlib_radar_formats.ipynb#GAMIC-HDF5`.
+    """
+
+    # check elevations
+    if wanted_elevations is None:
+        wanted_elevations = 'all'
+
+    # check wanted_moments
+    if wanted_moments is None:
+        wanted_moments = 'all'
+
+    # read the data from file
+    f = h5py.File(filename, 'r')
+
+    # placeholder for attributes and data
+    attrs = {}
+    vattrs = {}
+    data = {}
+
+    # check if GAMIC file and
+    try:
+        f['how'].attrs.get('software')
+    except KeyError:
+        print("WRADLIB: File is no GAMIC hdf5!")
+        raise
+
+    # get scan_type (PVOL or RHI)
+    scan_type = f['what'].attrs.get('object').decode()
+
+    # single or volume scan
+    if scan_type == 'PVOL':
+        # loop over 'main' hdf5 groups (how, scanX, what, where)
+        for n in list(f):
+            if 'scan' in n:
+                g = f[n]
+                sg1 = g['how']
+
+                # get scan elevation
+                el = sg1.attrs.get('elevation')
+                el = str(round(el, 2))
+
+                # try to read scan data and attrs
+                # if wanted_elevations are found
+                if (el in wanted_elevations) or (wanted_elevations == 'all'):
+                    sdata, sattrs = read_gamic_scan(scan=g,
+                                                    scan_type=scan_type,
+                                                    wanted_moments=wanted_moments)  # noqa
+                    if sdata:
+                        data[n.upper()] = sdata
+                    if sattrs:
+                        attrs[n.upper()] = sattrs
+
+    # single rhi scan
+    elif scan_type == 'RHI':
+        # loop over 'main' hdf5 groups (how, scanX, what, where)
+        for n in list(f):
+            if 'scan' in n:
+                g = f[n]
+                # try to read scan data and attrs
+                sdata, sattrs = read_gamic_scan(scan=g, scan_type=scan_type,
+                                                wanted_moments=wanted_moments)
+                if sdata:
+                    data[n.upper()] = sdata
+                if sattrs:
+                    attrs[n.upper()] = sattrs
+
+    # collect volume attributes if wanted data is available
+    if data:
+        vattrs['Latitude'] = f['where'].attrs.get('lat')
+        vattrs['Longitude'] = f['where'].attrs.get('lon')
+        vattrs['Height'] = f['where'].attrs.get('height')
+        # check whether its useful to implement that feature
+        # vattrs['sitecoords'] = (vattrs['Longitude'], vattrs['Latitude'],
+        #                         vattrs['Height'])
+        attrs['VOL'] = vattrs
+
+    f.close()
+
+    return data, attrs
+
+
+def to_hdf5(fpath, data, mode="w", metadata=None,
+            dataset="data", compression="gzip"):
+    """Quick storage of one <data> array and a <metadata> dict in an hdf5 file
+
+    This is more efficient than pickle, cPickle or numpy.save. The data is
+    stored in a subgroup named ``data`` (i.e. hdf5file["data").
+    See :meth:`~wradlib.io.from_hdf5` for retrieving stored data.
+
+    Parameters
+    ----------
+    fpath : string
+        path to the hdf5 file
+    data : :func:`numpy:numpy.array`
+    mode : string
+        file open mode, defaults to "w" (create, truncate if exists)
+    metadata : dict
+        dictionary of data's attributes
+    dataset : string
+        describing dataset
+    compression : string
+        h5py compression type {"gzip"|"szip"|"lzf"}, see h5py documentation
+        for details
+    """
+    f = h5py.File(fpath, mode=mode)
+    dset = f.create_dataset(dataset, data=data, compression=compression)
+    # store metadata
+    if metadata:
+        for key in metadata.keys():
+            dset.attrs[key] = metadata[key]
+    # close hdf5 file
+    f.close()
+
+
+def from_hdf5(fpath, dataset="data"):
+    """Loading data from hdf5 files that was stored by \
+    :meth:`~wradlib.io.to_hdf5`
+
+    Parameters
+    ----------
+    fpath : string
+        path to the hdf5 file
+    dataset : string
+        name of the Dataset in which the data is stored
+    """
+    f = h5py.File(fpath, mode="r")
+    # Check whether Dataset exists
+    if dataset not in f.keys():
+        print("Cannot read Dataset <%s> from hdf5 file <%s>" % (dataset, f))
+        f.close()
+        sys.exit()
+    data = np.array(f[dataset][:])
+    # get metadata
+    metadata = {}
+    for key in f[dataset].attrs.keys():
+        metadata[key] = f[dataset].attrs[key]
+    f.close()
+    return data, metadata
