+++ wradlib/io/misc.py
@@ -1,0 +1,102 @@
+#!/usr/bin/env python
+# Copyright (c) 2011-2017, wradlib developers.
+# Distributed under the MIT License. See LICENSE.txt for more info.
+
+"""
+Miscellaneous Data I/O
+^^^^^^^^^^^^^^^^^^^^^^
+.. autosummary::
+   :nosignatures:
+   :toctree: generated/
+
+   writePolygon2Text
+   to_pickle
+   from_pickle
+"""
+
+# standard libraries
+from __future__ import absolute_import
+
+try:
+    import cPickle as pickle
+except ImportError:
+    import pickle
+
+
+def _write_polygon2txt(f, idx, vertices):
+    f.write('%i %i\n' % idx)
+    for i, vert in enumerate(vertices):
+        f.write('%i ' % (i,))
+        f.write('%f %f %f %f\n' % tuple(vert))
+
+
+def writePolygon2Text(fname, polygons):
+    """Writes Polygons to a Text file which can be interpreted by ESRI \
+    ArcGIS's "Create Features from Text File (Samples)" tool.
+
+    This is (yet) only a convenience function with limited functionality.
+    E.g. interior rings are not yet supported.
+
+    Parameters
+    ----------
+    fname : string
+        name of the file to save the vertex data to
+    polygons : list of lists
+        list of polygon vertices.
+        Each vertex itself is a list of 3 coordinate values and an
+        additional value. The third coordinate and the fourth value may be nan.
+
+    Returns
+    -------
+    None
+
+    Note
+    ----
+    As Polygons are closed shapes, the first and the last vertex of each
+    polygon **must** be the same!
+
+    Examples
+    --------
+    Writes two triangle Polygons to a text file::
+        poly1 = [[0.,0.,0.,0.],[0.,1.,0.,1.],[1.,1.,0.,2.],[0.,0.,0.,0.]]
+        poly2 = [[0.,0.,0.,0.],[0.,1.,0.,1.],[1.,1.,0.,2.],[0.,0.,0.,0.]]
+        polygons = [poly1, poly2]
+        writePolygon2Text('polygons.txt', polygons)
+    The resulting text file will look like this::
+        Polygon
+        0 0
+        0 0.000000 0.000000 0.000000 0.000000
+        1 0.000000 1.000000 0.000000 1.000000
+        2 1.000000 1.000000 0.000000 2.000000
+        3 0.000000 0.000000 0.000000 0.000000
+        1 0
+        0 0.000000 0.000000 0.000000 0.000000
+        1 0.000000 1.000000 0.000000 1.000000
+        2 1.000000 1.000000 0.000000 2.000000
+        3 0.000000 0.000000 0.000000 0.000000
+        END
+    """
+    with open(fname, 'w') as f:
+        f.write('Polygon\n')
+        count = 0
+        for vertices in polygons:
+            _write_polygon2txt(f, (count, 0), vertices)
+            count += 1
+        f.write('END\n')
+
+
+def to_pickle(fpath, obj):
+    """Pickle object <obj> to file <fpath>
+    """
+    output = open(fpath, 'wb')
+    pickle.dump(obj, output)
+    output.close()
+
+
+def from_pickle(fpath):
+    """Return pickled object from file <fpath>
+    """
+    pkl_file = open(fpath, 'rb')
+    obj = pickle.load(pkl_file)
+    pkl_file.close()
+    return obj
